Add placeholder layout for PlannerView
Quick edit
